+++ lib/src/utils/par_sequence_n_executor.dart
@@ -1,36 +1,39 @@
 part of '../dart_either.dart';
 
-@internal
-class ParSequenceNExecutor<L, R> {
-  final Iterable<Future<Either<L, R>> Function()> functions;
-  final int? maxConcurrent;
+final class _ParSequenceNExecutor<L, R> {
+  final Iterable<Future<Either<L, R>> Function()> _functions;
+  final int? _maxConcurrent;
 
-  ParSequenceNExecutor(this.functions, this.maxConcurrent);
+  _ParSequenceNExecutor(this._functions, this._maxConcurrent);
 
   Future<Either<L, BuiltList<R>>> run() {
-    final mx = maxConcurrent;
+    final mx = _maxConcurrent;
 
     ControlError<L>? leftRaised;
 
-    final futureFunctions = functions.toList(growable: false);
+    final futureFunctions = _functions.toList(growable: false);
     final semaphore = mx != null ? Semaphore(mx) : null;
     final token = _Token();
 
     Future<R> Function() run(Future<Either<L, R>> Function() f) {
       return () {
-        // Check before execution of f.
+        // Reject a queued function before invoking it.
         if (leftRaised != null) {
           throw leftRaised!;
         }
         return Future.sync(f).then(
           (e) {
-            // Check after then the f is completed.
+            // Preserve the first Left observed while this function was running.
             if (leftRaised != null) {
               throw leftRaised!;
             }
 
             return e.fold(
-              ifLeft: (l) => throw ControlError<L>._(l, token),
+              ifLeft: (l) {
+                final error = ControlError<L>._(l, token);
+                leftRaised ??= error;
+                throw leftRaised!;
+              },
               ifRight: identity,
             );
           },
@@ -49,11 +52,8 @@ class ParSequenceNExecutor<L, R> {
     )
         .then((values) => Either<L, BuiltList<R>>.right(values.build()))
         .onError<ControlError<L>>(
-      (e, s) {
-        leftRaised = e;
-        return Left(e._value);
-      },
-      test: (e) => identical(e._token, token),
-    );
+          (e, s) => Left(e._value),
+          test: (e) => identical(e._token, token),
+        );
   }
 }